chore: Move DefaultDialog File
