+++ app/src/main/java/com/daily/dayo/presentation/fragment/post/PostOptionFragment.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.DialogFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.daily.dayo.databinding.FragmentPostOptionBinding
-import com.daily.dayo.common.DefaultDialogConfigure
+import com.daily.dayo.common.dialog.DefaultDialogConfigure
 import com.daily.dayo.common.autoCleared
 
 class PostOptionFragment : DialogFragment() {
