+++ app/src/main/java/com/daily/dayo/presentation/fragment/write/WriteImageOptionFragment.kt
@@ -19,7 +19,7 @@ import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import com.daily.dayo.BuildConfig
 import com.daily.dayo.R
-import com.daily.dayo.common.DefaultDialogConfigure
+import com.daily.dayo.common.dialog.DefaultDialogConfigure
 import com.daily.dayo.common.autoCleared
 import com.daily.dayo.databinding.FragmentWriteImageOptionBinding
 import com.daily.dayo.presentation.viewmodel.WriteViewModel
